fixed logout, timestamps

+++ src/main/resources/db/migration/V5__buildings.sql
@@ -1,7 +1,7 @@
 CREATE TABLE buildings (id INT AUTO_INCREMENT PRIMARY KEY,
                         type VARCHAR(45),
-                        level SMALLINT (4),
-                        hp SMALLINT (4),
+                        level INT (4),
+                        hp INT (4),
                         started_at BIGINT (10),
                         finished_at BIGINT (10) );
